+++ src/mini_agent/tools/__init__.py
@@ -1,6 +1,7 @@
 from mini_agent.tools.base import ToolRegistry, ToolExecutor, Tool
 from mini_agent.tools.calc import calculate_tool
 from mini_agent.tools.file import read_file_tool, write_file_tool
+from mini_agent.permission import PermissionGate
 
 registry = ToolRegistry()
 registry.register(calculate_tool)